Add animations story
