+++ .storybook/addons.js
@@ -1,2 +1,3 @@
 import '@storybook/addon-knobs/register';
 import '@storybook/addon-docs/register';
+import '@storybook/addon-actions/register';
