+++ .storybook/config.js
@@ -1,5 +1,6 @@
 import { configure, addParameters } from '@storybook/html';
 import { DocsPage, DocsContainer } from '@storybook/addon-docs/blocks';
+import '@storybook/addon-console';
 
 addParameters({
   docs: {
