+++ sandbox/index.html
@@ -9,7 +9,6 @@
 
     <ul>
       <li><a href="html/index.html">Sandbox Platformer</a></li>
-      <li><a href="tests/animation/animation.html">Animations</a></li>
       <li><a href="tests/gif/animatedGif.html">Animated Gif</a></li>
       <li><a href="tests/culling/culling.html">Sprite Culling</a></li>
       <li><a href="tests/culling/culling2.html">Zoom Culling</a></li>
