--- sandbox/tests/animation/animation.html
@@ -1,13 +1,0 @@
-﻿<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-    <title>Animation Test</title>
-</head>
-<body>
-    <canvas id="game"></canvas>
-    <script src="../../lib/excalibur.js"></script>
-    <script src="animation.js"></script>
-    <p>Press 'd' for debug mode</p>
-    <p>An animation with 3 frames should play, stopping on the third frame.</p>
-</body>
-</html>
