+++ src/stories/assets/animation.png
