+++ src/stories/globals.d.ts
@@ -1,0 +1,7 @@
+interface Story {
+  (): Element;
+  story?: {
+    decorators?: any;
+    parameters?: any;
+  };
+}
