+++ src/stories/utils.ts
@@ -1,4 +1,4 @@
-import { Engine, DisplayMode } from '../engine';
+import { DisplayMode, Engine, Input } from '../engine';
 
 /**
  * Helper to generate Storybook game engine instance
@@ -8,6 +8,12 @@ export const withEngine = (storyFn: (game: Engine) => void) => {
   return () => {
     const canvas = document.createElement('canvas');
     const game = new Engine({ canvasElement: canvas, displayMode: DisplayMode.FullScreen, suppressPlayButton: true });
+
+    game.input.keyboard.on('down', (keyDown?: Input.KeyEvent) => {
+      if (keyDown.key === Input.Keys.D) {
+        game.isDebug = !game.isDebug;
+      }
+    });
 
     storyFn(game);
